* release: (5311 commits)
  trigger redeployment
  Update: Grunt
  Update: Grunt string-replace
  Update: Version name update
  updated change log
  added change log
  PROD-5887 - pull the api changes
  PROD-5647 Fixed message read and unread action link on thread header when send new message
  Update: Grunt
  Update: Grunt string-replace
  Update: Version name update
  added change log
  PROD-5737 - setup the documentation link
  PROD-5737 - pull the api to platform
  PROD-5807 - Fixed group and standalone forum discussions notification not remove once delete forums discussion
  [PROD-5737] minify js
  [PROD-5737] Update condition to enable all domain rules options
  PROD-5805 - fix the issue with the widget hierarchy display
  PROD-5805: Update the logic to show the childs
  PROD-5805: Change function name and add doc Block for filter
  ...

# Conflicts:
#	src/bp-core/compatibility/bp-incompatible-plugins-helper.php

+++ src/bp-core/bp-core-tools-default-data.php
@@ -243,11 +243,13 @@ function bp_dd_delete_dummy_forum() {
 	/**
 	 * Delete Forums
 	 */
-	$forums = bp_get_option( 'bp_dd_imported_forum_ids' );
-	if ( ! empty( $forums ) ) {
-		foreach ( (array) $forums as $forum_id ) {
-			wp_delete_post( $forum_id );
-			bbp_delete_forum( $forum_id );
+	if ( bp_is_active( 'forums' ) ) {
+		$forums = bp_get_option( 'bp_dd_imported_forum_ids' );
+		if ( ! empty( $forums ) ) {
+			foreach ( (array) $forums as $forum_id ) {
+				wp_delete_post( $forum_id );
+				bbp_delete_forum( $forum_id );
+			}
 		}
 	}
 }
